+++ test/functional/rpc_blockchain.py
@@ -6,6 +6,7 @@
 """Test RPCs related to blockchainstate.
 
 Test the following RPCs:
+    - getblockchaininfo
     - gettxoutsetinfo
     - getdifficulty
     - getbestblockhash
@@ -61,6 +62,7 @@ class BlockchainTest(BitcoinTestFramework):
             'chainwork',
             'difficulty',
             'headers',
+            'initialblockdownload',
             'mediantime',
             'pruned',
             'size_on_disk',